+++ src/Spreads.Collections/Attributes.fs
@@ -1,11 +1,14 @@
 ﻿namespace Spreads
-// let related projects see implementation details
+
 [<assembly:System.Runtime.CompilerServices.InternalsVisibleTo("Spreads")>]
 [<assembly:System.Runtime.CompilerServices.InternalsVisibleTo("Spreads.Extensions")>]
-[<assembly:System.Runtime.CompilerServices.InternalsVisibleTo("SpreadsDB")>]
 [<assembly:System.Runtime.CompilerServices.InternalsVisibleTo("Spreads.Tests")>]
 [<assembly:System.Runtime.CompilerServices.InternalsVisibleTo("Spreads.Collections.Tests")>]
 [<assembly:System.Runtime.CompilerServices.InternalsVisibleTo("Spreads.Extensions.Tests")>]
 [<assembly:System.Runtime.CompilerServices.InternalsVisibleTo("Spreads.Tests.Profile")>]
+
+[<assembly:System.Runtime.CompilerServices.InternalsVisibleTo("SpreadsDB")>]
+[<assembly:System.Runtime.CompilerServices.InternalsVisibleTo("SpreadsDB.Tests")>]
+
 
 do ()