+++ src/Spreads.Extensions/Properties/AssemblyInfo.cs
@@ -4,11 +4,11 @@ using System.Reflection;
 [assembly: AssemblyTitleAttribute("Spreads.Extensions")]
 [assembly: AssemblyProductAttribute("Spreads")]
 [assembly: AssemblyDescriptionAttribute("Spreads")]
-[assembly: AssemblyVersionAttribute("0.1.8")]
-[assembly: AssemblyFileVersionAttribute("0.1.8")]
+[assembly: AssemblyVersionAttribute("0.1.9")]
+[assembly: AssemblyFileVersionAttribute("0.1.9")]
 [assembly: AssemblyCopyrightAttribute("(c) Victor Baybekov 2015")]
 namespace System {
     internal static class AssemblyVersionInformation {
-        internal const string Version = "0.1.8";
+        internal const string Version = "0.1.9";
     }
 }