+++ src/Spreads.Extensions/Properties/Attributes.cs
@@ -2,7 +2,8 @@
 
 [assembly: InternalsVisibleTo("Spreads")]
 [assembly: InternalsVisibleTo("Spreads.Tests")]
-[assembly: InternalsVisibleTo("SpreadsDB")]
-[assembly: InternalsVisibleTo("SpreadsDB.Tests")]
 [assembly: InternalsVisibleTo("Spreads.Tests.Profile")]
 [assembly: InternalsVisibleTo("Spreads.Extensions.Tests")]
+
+[assembly: InternalsVisibleTo("SpreadsDB")]
+[assembly: InternalsVisibleTo("SpreadsDB.Tests")]