+++ tests/Spreads.Collections.Tests/Profiler.cs
@@ -1,15 +1,18 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Spreads.Collections.Tests {
+    /// <summary>
+    /// Change the project type to a console app and run a profiler with Alt+F2 in VS
+    /// </summary>
     public static class Profiler {
         public static void Main() {
-            new ZipNTests().ContinuousZipIsCorrectByRandomCheck();
+            //new ZipNTests().ContinuousZipIsCorrectByRandomCheck();
             //var test = new MoveNextAsyncTests();
             //test.CouldReadSortedMapNewValuesWhileTheyAreAddedUsingCursor();
+
+
+            ////////////////////////////////////////
+            Console.WriteLine("Finished profiling");
             Console.ReadLine();
         }
     }